backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==11.0.2
python-dotenv==1.0.0
